requirements/pip-requirements.txt
pyaudio>=0.2.11
aftership>=0.2
feedparser>=5.2.1
pafy>=0.5.4
psutil>=5.4.8
urllib3>=1.24.1
google-api-python-client>=1.7.5
requests>=2.20.1
pyyaml>=3.13
spidev>=3.2
gpiozero>=1.4.1
numpy>=1.15.4
googletrans==3.1.0a0
cryptography==3.3.2
mock==3.0.5
oauth2client
soundfile
pyserial
